ECA crud methods now work with EntityTypes.
This is very handy when you wish to delete or create / update an EntityType,
but don't have the reference to which row it originated from.

An example use case is when one wishes to delete a row by clicking on some nested
imageView contained within. That imageViews onClick listener (bound in the
EntityCursorAdapters bindView method) will not receove a reference to which row
itself is contained in, but will have a reference to the current EntityType being
processed (which is provided through the bindView parameters).